requirements.txt
# ═════════════════════════════════════════════════════════════════════════════
# FastAPI & Web Framework
# ═════════════════════════════════════════════════════════════════════════════
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0


# ═════════════════════════════════════════════════════════════════════════════
# Google APIs
# ═════════════════════════════════════════════════════════════════════════════
google-generativeai>=0.3.0
google-cloud-texttospeech>=2.14.0
google-auth>=2.25.0


# ═════════════════════════════════════════════════════════════════════════════
# Document Processing
# ═════════════════════════════════════════════════════════════════════════════
PyPDF2>=3.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0


# ═════════════════════════════════════════════════════════════════════════════
# Database & ORM
# ═════════════════════════════════════════════════════════════════════════════
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0


# ═════════════════════════════════════════════════════════════════════════════
# Caching & Redis
# ═════════════════════════════════════════════════════════════════════════════
redis>=5.0.0


# ═════════════════════════════════════════════════════════════════════════════
# Video & Audio Processing
# ═════════════════════════════════════════════════════════════════════════════
Pillow>=10.0.0
pydub>=0.25.0


# ═════════════════════════════════════════════════════════════════════════════
# Machine Learning & Semantic Search
# ═════════════════════════════════════════════════════════════════════════════
numpy>=1.26.0
scikit-learn>=1.3.0
scipy>=1.11.0


# ═════════════════════════════════════════════════════════════════════════════
# AWS & Cloud
# ═════════════════════════════════════════════════════════════════════════════
boto3>=1.26.0


# ═════════════════════════════════════════════════════════════════════════════
# Configuration & Utilities
# ═════════════════════════════════════════════════════════════════════════════
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.25.0
cachetools>=5.3.0
pyyaml>=6.0.0


# ═════════════════════════════════════════════════════════════════════════════
# Development & Testing
# ═════════════════════════════════════════════════════════════════════════════
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
